services/policy-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis==5.0.4
pydantic==2.7.1
structlog==24.1.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
python-multipart==0.0.22
